Changed date format to display 24hours
The `format-date` function doc gives `15:23:35s` as an example but this will never be as the format for the hour is `hh` (Hour in am/pm (1-12)). Moreover there's no way to distinguish whether the formatted date is AM or PM. This tiny tweak gives a clear indication of what time of day is represented.